Add support for classifications and showing posts
